JobDefinitionParser logic moved to JobDefinitionFactory
