+++ src/JobDefinitionFactory.php
@@ -9,9 +9,12 @@ use Keboola\ConfigurationVariablesResolver\VariablesResolver;
 use Keboola\DockerBundle\Docker\Component;
 use Keboola\DockerBundle\Docker\JobDefinition;
 use Keboola\DockerBundle\Exception\ApplicationException;
+use Keboola\DockerBundle\Exception\UserException;
 use Keboola\JobQueueInternalClient\JobFactory\JobInterface;
 use Keboola\JobQueueInternalClient\JobFactory\ObjectEncryptor\JobObjectEncryptor;
 use Keboola\PermissionChecker\BranchType;
+use Keboola\StorageApi\ClientException;
+use Keboola\StorageApi\Components;
 use Keboola\StorageApi\DevBranches;
 use Keboola\StorageApiBranch\ClientWrapper;
 use Keboola\VaultApiClient\Variables\VariablesApiClient;
@@ -20,7 +23,7 @@ use Psr\Log\LoggerInterface;
 class JobDefinitionFactory
 {
     public function __construct(
-        private readonly JobDefinitionParser $jobDefinitionParser,
+        private readonly DockerBundleJobDefinitionParser $jobDefinitionParser,
         private readonly JobObjectEncryptor $objectEncryptor,
         private readonly VariablesApiClient $variablesApiClient,
         private readonly LoggerInterface $logger,
@@ -35,7 +38,7 @@ class JobDefinitionFactory
         JobInterface $job,
         ClientWrapper $clientWrapper,
     ): array {
-        $jobDefinitions = $this->jobDefinitionParser->createJobDefinitionsForJob(
+        $jobDefinitions = $this->createJobDefinitionsForJob(
             $clientWrapper,
             $component,
             $job,
@@ -54,6 +57,59 @@ class JobDefinitionFactory
         );
 
         return $jobDefinitions;
+    }
+
+    /**
+     * @return JobDefinition[]
+     */
+    public function createJobDefinitionsForJob(
+        ClientWrapper $clientWrapper,
+        Component $component,
+        JobInterface $job,
+    ): array {
+        if ($component->blockBranchJobs() && $clientWrapper->hasBranch()) {
+            throw new UserException('This component cannot be run in a development branch.');
+        }
+
+        if ($job->getConfigData()) {
+            $configData = $job->getConfigData();
+            $configData = $this->extendComponentConfigWithBackend($configData, $job);
+
+            $jobDefinition = $this->jobDefinitionParser->parseConfigData(
+                $component,
+                $configData,
+                $job->getConfigId(),
+                ($job->getBranchType() ?? BranchType::DEFAULT)->value,
+            );
+            return [$jobDefinition];
+        }
+
+        try {
+            $components = new Components($clientWrapper->getBranchClientIfAvailable());
+            $configuration = $components->getConfiguration($job->getComponentId(), $job->getConfigId());
+            /** @var array $configuration */
+
+            if (!$clientWrapper->getClientOptionsReadOnly()->useBranchStorage()) {
+                $this->checkUnsafeConfiguration(
+                    $component,
+                    $configuration,
+                    $job->getBranchType() ?? BranchType::DEV
+                );
+            }
+        } catch (ClientException $e) {
+            throw new UserException($e->getMessage(), $e);
+        }
+
+        $configuration['configuration'] = $this->extendComponentConfigWithBackend(
+            $configuration['configuration'] ?? [],
+            $job,
+        );
+
+        return $this->jobDefinitionParser->parseConfig(
+            $component,
+            $configuration,
+            ($job->getBranchType() ?? BranchType::DEFAULT)->value,
+        );
     }
 
     /**
@@ -132,5 +188,30 @@ class JobDefinitionFactory
             ),
             $jobDefinitions,
         );
+    }
+
+    private function extendComponentConfigWithBackend(array $config, JobInterface $job): array
+    {
+        $backend = $job->getBackend();
+
+        if ($backend->getType() !== null) {
+            $config['runtime']['backend']['type'] = $backend->getType();
+        }
+        if ($backend->getContext() !== null) {
+            $config['runtime']['backend']['context'] = $backend->getContext();
+        }
+
+        return $config;
+    }
+
+    private function checkUnsafeConfiguration(Component $component, array $configuration, BranchType $branchType): void
+    {
+        if ($component->branchConfigurationsAreUnsafe() && $branchType === BranchType::DEV) {
+            if (empty($configuration['configuration']['runtime']['safe'])) {
+                throw new UserException(
+                    'It is not safe to run this configuration in a development branch. Please review the configuration.'
+                );
+            }
+        }
     }
 }
